+++ backend/config/security.php
@@ -48,7 +48,8 @@ return [
     '/api/user/character/find-by'        => [Role::USER_ADMIN, Role::USER_MANAGER, Role::GROUP_MANAGER, Role::WATCHLIST],
     '/api/user/character/{id}/update'    => [Role::USER],
 
-    '/api/user/corporation/tracked-corporations'       => [Role::TRACKING, Role::TRACKING_ADMIN],
+    '/api/user/corporation/tracked-corporations'       => [Role::TRACKING],
+    '/api/user/corporation/all-tracked-corporations'   => [Role::TRACKING_ADMIN],
     '/api/user/corporation/{id}/members'               => [Role::TRACKING],
     '/api/user/corporation/{id}/tracking-director'     => [Role::TRACKING_ADMIN],
     '/api/user/corporation/{id}/get-groups-tracking'   => [Role::TRACKING_ADMIN],